+++ packages/gatsby-plugin-mdx/gatsby/on-create-node.js
@@ -4,8 +4,9 @@ const babel = require(`@babel/core`)
 const { createContentDigest } = require(`gatsby-core-utils`)
 
 const defaultOptions = require(`../utils/default-options`)
-const createMDXNodeWithScope = require(`../utils/mdx-node-with-scope`)
+const createMDXNode = require(`../utils/create-mdx-node`)
 const { MDX_SCOPES_LOCATION } = require(`../constants`)
+const { findImports } = require(`../utils/gen-mdx`)
 
 const contentDigest = val => createContentDigest(val)
 
@@ -45,14 +46,18 @@ module.exports = async (
 
   const content = await loadNodeContent(node)
 
-  const {
-    mdxNode,
-    scopeIdentifiers,
-    scopeImports,
-  } = await createMDXNodeWithScope({
+  const mdxNode = await createMDXNode({
     id: createNodeId(`${node.id} >>> Mdx`),
     node,
     content,
+  })
+
+  createNode(mdxNode)
+  createParentChildLink({ parent: node, child: mdxNode })
+
+  // write scope files into .cache for later consumption
+  const { scopeImports, scopeIdentifiers } = await findImports({
+    node: mdxNode,
     getNode,
     getNodes,
     reporter,
@@ -64,10 +69,6 @@ module.exports = async (
     createNodeId,
     ...helpers,
   })
-
-  createNode(mdxNode)
-  createParentChildLink({ parent: node, child: mdxNode })
-
   await cacheScope({
     cache,
     scopeIdentifiers,