+++ packages/gatsby-plugin-offline/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.2.22](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-offline@3.2.21...gatsby-plugin-offline@3.2.22) (2020-08-05)
+
+**Note:** Version bump only for package gatsby-plugin-offline
+
 ## [3.2.21](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-offline@3.2.20...gatsby-plugin-offline@3.2.21) (2020-07-24)
 
 **Note:** Version bump only for package gatsby-plugin-offline