+++ packages/gatsby-plugin-page-creator/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.20](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-page-creator@2.3.19...gatsby-plugin-page-creator@2.3.20) (2020-08-05)
+
+**Note:** Version bump only for package gatsby-plugin-page-creator
+
 ## [2.3.19](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-page-creator@2.3.18...gatsby-plugin-page-creator@2.3.19) (2020-07-24)
 
 **Note:** Version bump only for package gatsby-plugin-page-creator